fix: raise exception if nothing has been sent
For better error monitoring, the upload pack failure will appear on
"Response stats" log message: upload_pack_status="error" instead of
"direct" default value.
So we can check if we didn't serve the client because of
upload_pack_status=="error". If not this should not happen, we raise
an exception.

Signed-off-by: Sebastien MICHEL <sebastien.t.michel@renault.com>