fix: move cleaner after serving file
In the case PACK_CACHE_SIZE_GB is lower than the current generated pack
size, the cache_cleaner will delete the pack file before the main thread
serving it.
To prevent this case, we launch the cache_cleaner thread after serving
the pack, then after it can be deleted.
Calling cache_cleaner after serving the pack doesn't increase the
frequence of cache_cleaner call, because if we arrive there, it's
because the pack has just been created.

Signed-off-by: Sebastien MICHEL <sebastien.t.michel@renault.com>