fix: don't stream data if full body is buffered
We hit an odd case in our installation where a cache on our enterprise GitHub
instance started to fail requests coming from git_cdn but not from any other
sources. After a lot of tracing I tracked it down to a combination of two things:

  - The `Transfer-Encoding` header is set to `chunked`
  - The body is empty

In git_cdn.py:fix_response_headers the Transfer-Encoding is supposed to be
removed from the proxied request. However, this doesn't work fully because if
the `data` parameter to `ClientSession.request` is an instance of
`StreamReader` the `Transfer-Encoding: chunked` header is automatically
re-added.

The fix I've implemented here is to check if the entire request body has
already been read into memory and if it has, just to pass it along as a bytes
object instead of a stream. In particular this addresses the empty body case as
0 bytes will always be ready to go.

Signed-off-by: Nick Stevens <nick@bitcurry.com>