fix: handle git clone/fetch of huge git repo
When doing git clone/fetch of big repo/big pack files,
the communication with the server can be stopped by someone in the middle,
with a 504 (Gateway Timeout Error) error result.

This issue is not seen when doing the git clone/fetch from command line.
The reason seems linked to the '--progress' option.
It is enabled by default when execution is done from a terminal.

Enabling the option with subprocess exec is fixing the disconnection issue.
Having traffic/progress on the connection, keep this one opened.
W/o it, the connection may be seen as stalled and terminated.

Signed-off-by: Sebastien Fusilier <sebastien.fusilier@renault.com>